feat: add exercise on model evaluation
The note about the decision tree model being too specific and possibly
overfitting is left hanging a little. If participants were to evaluate
the model on the test set, they can easily confirm this. The course
never really gave any concrete examples of overfitting, so an exercise
here might be nice for consolodating this knowledge